pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "python-doctr"
description = "Document Text Recognition (docTR): deep Learning for high-performance OCR on documents."
authors = [{name = "Mindee", email = "contact@mindee.com"}]
maintainers = [
    {name = "François-Guillaume Fernandez"},
    {name = "Charles Gaillard"},
    {name = "Olivier Dulcy"},
    {name = "Felix Dittrich"},
]
readme = "README.md"
requires-python = ">=3.9.0,<4"
license = {file = "LICENSE"}
keywords=["OCR", "deep learning", "computer vision", "tensorflow", "pytorch", "text detection", "text recognition"]
classifiers=[
        "Development Status :: 4 - Beta",
        "Intended Audience :: Developers",
        "Intended Audience :: Education",
        "Intended Audience :: Science/Research",
        "License :: OSI Approved :: Apache Software License",
        "Natural Language :: English",
        "Operating System :: OS Independent",
        "Programming Language :: Python :: 3",
        "Programming Language :: Python :: 3.9",
        "Programming Language :: Python :: 3.10",
        "Programming Language :: Python :: 3.11",
        "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dynamic = ["version"]
dependencies = [
    "importlib_metadata",
    # For proper typing, mypy needs numpy>=1.20.0 (cf. https://github.com/numpy/numpy/pull/16515)
    # Additional typing support is brought by numpy>=1.22.4, but core build sticks to >=1.16.0
    "numpy>=1.16.0,<2.0.0",
    "scipy>=1.4.0,<2.0.0",
    "h5py>=3.1.0,<4.0.0",
    "opencv-python>=4.5.0,<5.0.0",
    "pypdfium2>=4.0.0,<5.0.0",
    "pyclipper>=1.2.0,<2.0.0",
    "shapely>=1.6.0,<3.0.0",
    "langdetect>=1.0.9,<2.0.0",
    "rapidfuzz>=3.0.0,<4.0.0",
    "matplotlib>=3.1.0",
    "weasyprint>=55.0",
    "Pillow>=9.2.0",
    "defusedxml>=0.7.0",
    "mplcursors>=0.3",
    "unidecode>=1.0.0",
    "tqdm>=4.30.0",
    "huggingface-hub>=0.5.0",
]

[project.optional-dependencies]
tf = [
    "tensorflow>=2.11.0,<3.0.0",  # cf. https://github.com/mindee/doctr/pull/1182
    "tf2onnx>=1.15.1,<2.0.0",  # cf.https://github.com/onnx/tensorflow-onnx/releases/tag/v1.15.1
    "onnx>=1.12.0,!=1.15.0,<3.0.0",  # TODO: remove this line once tf2onnx supports onnx>=1.15.0
]
torch = [
    "torch>=1.12.0,<3.0.0",
    "torchvision>=0.13.0",
    "onnx>=1.12.0,<3.0.0",
]
testing = [
    "pytest>=5.3.2",
    "coverage[toml]>=4.5.4",
    "hdf5storage>=0.1.18",
    "onnxruntime>=1.11.0",
    "requests>=2.20.0",
    "psutil>=5.9.5"
]
quality = [
    "ruff>=0.1.5",
    "mypy>=0.812",
    "pre-commit>=2.17.0",
]
docs = [
    "sphinx>=3.0.0,!=3.5.0",
    "sphinxemoji>=0.1.8",
    "sphinx-copybutton>=0.3.1",
    "docutils<0.21",
    "recommonmark>=0.7.1",
    "sphinx-markdown-tables>=0.0.15",
    "sphinx-tabs>=3.3.0",
    "furo>=2022.3.4",
]
dev = [
    # Tensorflow
    "tensorflow>=2.11.0,<3.0.0",  # cf. https://github.com/mindee/doctr/pull/1182
    "tf2onnx>=1.15.1,<2.0.0",  # cf.https://github.com/onnx/tensorflow-onnx/releases/tag/v1.15.1
    # PyTorch
    "torch>=1.12.0,<3.0.0",
    "torchvision>=0.13.0",
    # TODO: remove the constraint once tf2onnx supports onnx>=1.15.0
    # TODO: cf. https://github.com/onnx/tensorflow-onnx/pull/2252
    "onnx>=1.12.0,!=1.15.0,<3.0.0",
    # Testing
    "pytest>=5.3.2",
    "coverage[toml]>=4.5.4",
    "hdf5storage>=0.1.18",
    "onnxruntime>=1.11.0",
    "requests>=2.20.0",
    "psutil>=5.9.5",
    # Quality
    "ruff>=0.1.5",
    "mypy>=0.812",
    "pre-commit>=2.17.0",
    # Docs
    "sphinx>=3.0.0,!=3.5.0",
    "sphinxemoji>=0.1.8",
    "sphinx-copybutton>=0.3.1",
    "docutils<0.21",
    "recommonmark>=0.7.1",
    "sphinx-markdown-tables>=0.0.15",
    "sphinx-tabs>=3.3.0",
    "furo>=2022.3.4",
]

[project.urls]
documentation = "https://mindee.github.io/doctr"
repository = "https://github.com/mindee/doctr"
tracker = "https://github.com/mindee/doctr/issues"
changelog = "https://github.com/mindee/doctr/latest/changelog.html"

[tool.setuptools]
zip-safe = true

[tool.setuptools.packages.find]
exclude = ["api*", "demo*", "docs*", "notebooks*", "references*", "scripts*", "tests*"]

[tool.mypy]
files = "doctr/"
show_error_codes = true
pretty = true
warn_unused_ignores = true
warn_redundant_casts = true
no_implicit_optional = true
check_untyped_defs = true
implicit_reexport = false

[[tool.mypy.overrides]]
module = [
	"tensorflow.*",
	"torchvision.*",
	"PIL.*",
	"tqdm.*",
	"scipy.*",
	"cv2.*",
	"h5py.*",
	"matplotlib.*",
	"pyclipper.*",
	"shapely.*",
	"tf2onnx.*",
	"mplcursors.*",
	"defusedxml.*",
	"weasyprint.*",
	"huggingface_hub.*",
	"pypdfium2.*",
	"rapidfuzz.*",
	"langdetect.*",
]
ignore_missing_imports = true

[tool.ruff]
select = [
    # https://docs.astral.sh/ruff/rules/
    "E", "W", "F", "I", "N", "Q", "C4",
    "D101", "D103", "D201","D202","D207","D208","D214","D215","D300","D301","D417", "D419",  # pydocstyle
]
ignore = ["E402", "F403", "E731", "N812", "N817", "C408", "PT"]
exclude = [".git", "venv*", "build"]
line-length = 120
target-version = "py39"

[tool.ruff.isort]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]
known-first-party = ["doctr", "tests", "scripts", "references", "demo", "docs", "api"]
known-third-party = ["tensorflow", "torch", "torchvision", "wandb", "fastprogress"]

[tool.ruff.per-file-ignores]
"**/__init__.py" = ["F401"]
"doctr/models/**.py" = ["N806"]
"doctr/datasets/**.py" = ["N806"]
"tests/**.py" = ["D"]
"scripts/**.py" = ["D"]
"docs/**.py" = ["D"]
"references/**.py" = ["D"]
"api/**.py" = ["D"]
"demo/**.py" = ["D"]
".github/**.py" = ["D"]


[tool.ruff.flake8-quotes]
docstring-quotes = "double"

[tool.coverage.run]
source = ["doctr"]
